requirements.txt
pulsar-client
fastavro
